+++ apps/rahat-ui/src/sections/communications/add/add-audiences.tsx
@@ -1,6 +1,10 @@
 'use client';
 
-import { useProjectList } from '@rahat-ui/query';
+import {
+  useBeneficiaryPii,
+  usePagination,
+  useProjectList,
+} from '@rahat-ui/query';
 import {
   FormField,
   FormItem,
@@ -23,34 +27,52 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@rahat-ui/shadcn/src/components/ui/select';
+import { useCreateAudience } from '@rumsan/communication-query';
 
 import { flexRender } from '@tanstack/react-table';
+import CustomPagination from 'apps/rahat-ui/src/components/customPagination';
 import React, { FC } from 'react';
 import { UseFormReturn } from 'react-hook-form';
 import { z } from 'zod';
+import { useAudienceColumns } from './use-audience-columns';
+import { useAudienceTable } from './use-audience-table';
 
 type AddAudienceProps = {
-  table: any;
   form: UseFormReturn<z.infer<any>>;
-  columns: any;
-  filters: any;
-  setFilters: any;
+
   globalFilter: any;
   setGlobalFilter: any;
   selectedRows: Array<any>;
+  audienceData: any;
+  setSelectedRows: any;
 };
 
 const AddAudience: FC<AddAudienceProps> = ({
-  table,
   form,
-  columns,
-  filters,
-  setFilters,
   globalFilter,
   setGlobalFilter,
   selectedRows,
+  audienceData,
+  setSelectedRows,
 }) => {
+  const [columnVisibility, setColumnVisibility] = React.useState({});
+  const [rowSelection, setRowSelection] = React.useState({});
+
   const projectsList = useProjectList({});
+
+  const {
+    pagination,
+    filters,
+    setFilters,
+    setNextPage,
+    setPrevPage,
+    setPerPage,
+  } = usePagination();
+  const { data: beneficiaryData } = useBeneficiaryPii({
+    ...pagination,
+    ...filters,
+  });
+  const createAudience = useCreateAudience();
 
   const filterBenByAudience = React.useCallback(
     (id: string) => {
@@ -62,13 +84,42 @@ const AddAudience: FC<AddAudienceProps> = ({
     },
     [filters, setFilters],
   );
+  const columns = useAudienceColumns(
+    beneficiaryData,
+    selectedRows,
+    audienceData,
+    createAudience,
+    setSelectedRows,
+  );
+
+  const tableData = React.useMemo(() => {
+    return (
+      beneficiaryData &&
+      beneficiaryData?.data?.map((item: any) => ({
+        name: item?.name,
+        id: item?.beneficiaryId,
+        phone: item?.phone,
+      }))
+    );
+  }, [beneficiaryData]);
+
+  const table = useAudienceTable({
+    columnVisibility,
+    columns,
+    globalFilter,
+    rowSelection,
+    setColumnVisibility,
+    setGlobalFilter,
+    setRowSelection,
+    tableData,
+  });
 
   return (
     <>
       {/* header area start  */}
       <div className="flex items-center gap-2 pb-2">
         <Input
-          placeholder="Filter audience..."
+          placeholder="Filter audiences"
           value={globalFilter ?? ''}
           onChange={(value) => {
             setGlobalFilter(value.target.value);
@@ -76,7 +127,7 @@ const AddAudience: FC<AddAudienceProps> = ({
           className="max-w-sm"
         />
         <Select onValueChange={filterBenByAudience}>
-          <SelectTrigger className='max-w-32'>
+          <SelectTrigger className="max-w-32">
             <SelectValue placeholder="Projects" />
           </SelectTrigger>
           <SelectContent>
@@ -91,7 +142,15 @@ const AddAudience: FC<AddAudienceProps> = ({
               })}
           </SelectContent>
         </Select>
-        <div className={`border rounded px-3 py-2 h-10 text-sm ${selectedRows.length ? 'bg-primary text-white font-medium' : 'bg-card'}`}>{selectedRows.length} - Audience selected</div>
+        <div
+          className={`border rounded px-3 py-2 h-10 text-sm ${
+            selectedRows.length
+              ? 'bg-primary text-white font-medium'
+              : 'bg-card'
+          }`}
+        >
+          {selectedRows.length} - Audience selected
+        </div>
       </div>
       {/* header area end  */}
       <FormField
@@ -101,7 +160,7 @@ const AddAudience: FC<AddAudienceProps> = ({
           <FormItem>
             <div className="rounded border mb-2 bg-card">
               <Table>
-                <ScrollArea className='h-[calc(100vh-376px)]'>
+                <ScrollArea className="h-[calc(100vh-376px)]">
                   <TableHeader>
                     {table.getHeaderGroups().map((headerGroup) => (
                       <TableRow key={headerGroup.id}>
@@ -111,9 +170,9 @@ const AddAudience: FC<AddAudienceProps> = ({
                               {header.isPlaceholder
                                 ? null
                                 : flexRender(
-                                  header.column.columnDef.header,
-                                  header.getContext(),
-                                )}
+                                    header.column.columnDef.header,
+                                    header.getContext(),
+                                  )}
                             </TableHead>
                           );
                         })}
@@ -150,6 +209,20 @@ const AddAudience: FC<AddAudienceProps> = ({
                   </TableBody>
                 </ScrollArea>
               </Table>
+              <CustomPagination
+                meta={
+                  beneficiaryData?.response?.meta || {
+                    total: 0,
+                    currentPage: 0,
+                  }
+                }
+                handleNextPage={setNextPage}
+                handlePrevPage={setPrevPage}
+                handlePageSizeChange={setPerPage}
+                currentPage={pagination.page}
+                perPage={pagination.perPage}
+                total={beneficiaryData?.response?.meta?.lastPage || 0}
+              />
             </div>
             <FormMessage />
           </FormItem>
